+++ Rockxy/Views/Inspector/AIInspectorView.swift
@@ -42,7 +42,7 @@ struct AIInspectorView: View {
                 )
             }
         }
-        .task(id: transaction.id) {
+        .task(id: transaction.inspectionKey) {
             await loadInspection()
         }
     }